Add frame

+++ solver/models/vector_2.py
@@ -2,5 +2,5 @@ from dataclasses import dataclass
 
 @dataclass
 class Vector2:
-	x: int
-	y: int
+	x: float
+	y: float
